# docs/api_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Equity Lab",
    "description": "Institutional Indian Equity Research, Options Arbitrage, and Return Probability Engine",
    "version": "0.0.9"
  },
  "paths": {
...
    "/api/v1/query": {
      "post": {
        "tags": [
          "AI Strategy Assistant"
        ],
        "summary": "Handle Ai Query",
        "description": "Processes research queries through verified LLM or deterministic analytical engine.",
        "operationId": "handle_ai_query_api_v1_query_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/companies": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Upsert Company",
        "operationId": "upsert_company_api_v1_data_companies_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/financial-observations": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Financial Observation",
        "operationId": "add_financial_observation_api_v1_data_financial_observations_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialObservationIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialObservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/business-events": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Business Event",
        "operationId": "add_business_event_api_v1_data_business_events_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessEventIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/corporate-actions": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Corporate Action",
        "operationId": "add_corporate_action_api_v1_data_corporate_actions_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorporateActionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorporateActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/ownership-snapshots": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Ownership Snapshot",
        "operationId": "add_ownership_snapshot_api_v1_data_ownership_snapshots_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipSnapshotIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/document-metadata": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Document Metadata",
        "operationId": "add_document_metadata_api_v1_data_document_metadata_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentMetadataIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/market-snapshots": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Add Market Daily Snapshot",
        "operationId": "add_market_daily_snapshot_api_v1_data_market_snapshots_post",
        "parameters": [
          {
            "name": "x-data-write-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Data-Write-Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDailySnapshotIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDailySnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/companies/{symbol}/timeline": {
      "get": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Get Company Timeline",
        "operationId": "get_company_timeline_api_v1_data_companies__symbol__timeline_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/lifecycle/{symbol}": {
      "get": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Get Lifecycle State",
        "operationId": "get_lifecycle_state_api_v1_data_lifecycle__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/thesis/{symbol}": {
      "get": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Get Thesis State",
        "operationId": "get_thesis_state_api_v1_data_thesis__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/alerts": {
      "get": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Get Alerts",
        "operationId": "get_alerts_api_v1_data_alerts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/custom-screen": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Run Custom Screen",
        "description": "Execute custom Screener.in query or preset, with optional 27-factor Multibagger Brain ranking.",
        "operationId": "run_custom_screen_api_v1_data_custom_screen_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/causal-inference": {
      "post": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Run Causal Inference",
        "operationId": "run_causal_inference_api_v1_data_causal_inference_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/thesis-tracker/{symbol}": {
      "get": {
        "tags": [
          "Point-in-Time Research Data"
        ],
        "summary": "Get Thesis Tracker",
        "operationId": "get_thesis_tracker_api_v1_data_thesis_tracker__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyUpsertRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Symbol"
          },
          "legal_name": {
            "type": "string",
            "maxLength": 250,
            "minLength": 2,
            "title": "Legal Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "legal_name"
        ],
        "title": "CompanyUpsertRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Query",
            "description": "Research prompt or query"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode",
            "description": "Query mode: Quick, Research, Technical, Fundamental",
            "default": "Quick"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Original research prompt"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Query mode used",
            "default": "Quick"
          },
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "Narrative response or deterministic fallback"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "LLM provider or deterministic engine identifier"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaHeader",
            "description": "Metadata header with source and timestamps"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "Legal disclaimer for the response"
          }
        },
        "type": "object",
        "required": [
          "query",
          "reply",
          "provider",
          "meta",
          "disclaimer"
        ],
        "title": "QueryResponse",
        "description": "Response model for AI query endpoint.\n\nMirrors the fields returned by `process_llm_query`.",
        "example": {
          "disclaimer": "AI research assistant readout for educational/research purposes. Not investment advice.",
          "meta": {
            "as_of": "2026-08-16T09:00:00+05:30",
            "limitations": [],
            "market_data_type": "realtime",
            "retrieved_at": "2026-08-16T09:00:05+05:30",
            "source": "IERL Market Data",
            "stale": false
          },
          "mode": "Quick",
          "provider": "Google Gemini 1.5 Flash",
          "query": "Show revenue growth for RELIANCE",
          "reply": "Revenue grew 12% YoY over the last quarter..."
        }
      },
...
    }
  }
}